pyproject.toml
[project]
name = "extracao-infos-pdf-darf-flask-web-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-migrate>=4.0.0",
    "flask-sqlalchemy>=3.1.0",
    "gunicorn>=21.2.0",
    "msal>=1.34.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.1",
    "pyqt6>=6.6.0",
    "rapidocr-onnxruntime>=1.3.0",
    "sqlalchemy>=2.0.0",
    "werkzeug>=3.1.3",
    "pywebview>=5.0.0",
    "waitress>=3.0.0",
    "onnxruntime>=1.7.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.json"]
